+++ client/cluster-client/src/main/kotlin/uk/ac/ebi/biostd/client/cluster/api/LocalClusterClient.kt
@@ -5,14 +5,14 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import uk.ac.ebi.biostd.client.cluster.model.Job
 import uk.ac.ebi.biostd.client.cluster.model.JobSpec
-import java.util.UUID
+import java.io.File
 
-class LocalClusterClient : ClusterOperations {
+class LocalClusterClient : ClusterClient {
     override suspend fun triggerJob(jobSpec: JobSpec): Try<Job> {
         return withContext(Dispatchers.IO) {
-            Runtime.getRuntime().exec(jobSpec.command)
-            val jobId = UUID.randomUUID().toString()
-            Try.just(Job(jobId, LOCAL_QUEUE, ""))
+            val logFile = File.createTempFile("prefix-", "-suffix");
+            val jobId = executeProcess(listOf(jobSpec.command), logFile)
+            Try.just(Job(jobId.toString(), LOCAL_QUEUE, logFile.absolutePath))
         }
     }
 
@@ -22,5 +22,13 @@ class LocalClusterClient : ClusterOperations {
 
     companion object {
         internal const val LOCAL_QUEUE = "local"
+    }
+
+    private fun executeProcess(params: List<String>, logFile: File): Long {
+        val processBuilder = ProcessBuilder(params)
+        processBuilder.redirectOutput(logFile)
+        val process = processBuilder.start()
+        process.waitFor()
+        return process.pid()
     }
 }
