+++ scheduler/scheduler/src/main/kotlin/uk/ac/ebi/scheduler/common/config/SchedulerConfig.kt
@@ -7,7 +7,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.scheduling.annotation.EnableScheduling
 import uk.ac.ebi.biostd.client.cluster.api.ClusterClient
-import uk.ac.ebi.biostd.client.cluster.api.ClusterOperations
+import uk.ac.ebi.biostd.client.cluster.api.LsfClusterClient
 import uk.ac.ebi.scheduler.common.properties.AppProperties
 import uk.ac.ebi.scheduler.pmc.exporter.api.ExporterProperties
 import uk.ac.ebi.scheduler.pmc.exporter.domain.ExporterTrigger
@@ -26,7 +26,7 @@ internal class SchedulerConfig {
     @Bean
     fun clusterOperations(
         appProperties: AppProperties,
-    ): ClusterOperations = ClusterClient.create(
+    ): ClusterClient = LsfClusterClient.create(
         appProperties.cluster.sshKey,
         appProperties.cluster.server,
         appProperties.cluster.logsPath,
@@ -34,32 +34,32 @@ internal class SchedulerConfig {
 
     @Bean
     fun loaderService(
-        clusterOperations: ClusterOperations,
+        clusterClient: ClusterClient,
         properties: PmcProcessorProp,
         appProperties: AppProperties,
         @Qualifier("pmcNotificationsSender") pmcNotificationsSender: NotificationsSender,
-    ): PmcLoaderService = PmcLoaderService(clusterOperations, properties, appProperties, pmcNotificationsSender)
+    ): PmcLoaderService = PmcLoaderService(clusterClient, properties, appProperties, pmcNotificationsSender)
 
     @Bean
     fun submissionReleaserTrigger(
         appProperties: AppProperties,
-        clusterOperations: ClusterOperations,
+        clusterClient: ClusterClient,
         releaserProperties: SubmissionReleaserProperties,
         @Qualifier("schedulerNotificationsSender") schedulerNotificationsSender: NotificationsSender,
     ): SubmissionReleaserTrigger =
-        SubmissionReleaserTrigger(appProperties, releaserProperties, clusterOperations, schedulerNotificationsSender)
+        SubmissionReleaserTrigger(appProperties, releaserProperties, clusterClient, schedulerNotificationsSender)
 
     @Bean
     fun exporterTrigger(
         appProperties: AppProperties,
-        clusterOperations: ClusterOperations,
+        clusterClient: ClusterClient,
         exporterProperties: ExporterProperties,
         @Qualifier("pmcNotificationsSender") pmcNotificationsSender: NotificationsSender,
         @Qualifier("schedulerNotificationsSender") schedulerNotificationsSender: NotificationsSender,
     ): ExporterTrigger = ExporterTrigger(
         appProperties,
         exporterProperties,
-        clusterOperations,
+        clusterClient,
         pmcNotificationsSender,
         schedulerNotificationsSender,
     )
@@ -68,12 +68,12 @@ internal class SchedulerConfig {
     fun statsReporterTrigger(
         appProperties: AppProperties,
         properties: StatsReporterProperties,
-        clusterOperations: ClusterOperations,
+        clusterClient: ClusterClient,
         schedulerNotificationsSender: NotificationsSender,
     ): StatsReporterTrigger = StatsReporterTrigger(
         appProperties,
         properties,
-        clusterOperations,
+        clusterClient,
         schedulerNotificationsSender,
     )
